+++ numpy_fracadf/__init__.py
@@ -1,1 +1,2 @@
 from .fracadf1 import fracadf1
+from .fracadf2 import fracadf2